requirements.txt
aiokafka==0.12.0
aiobotocore[boto3]==3.0.0
fastapi>=0.124.4
pydantic>=2.12.5
python-multipart>=0.0.20
uvicorn[standard]>=0.38.0
pyjwt[crypto]==2.10.1
